refactoring tracing module
