+++ integrations/agent_pack/src/haystack_integrations/tracing/agent_pack/tracer.py
@@ -2,12 +2,10 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from _thread import LockType
 from collections.abc import Iterator
 from contextlib import contextmanager
 from contextvars import ContextVar
-from dataclasses import dataclass, field
-from threading import Lock
+from dataclasses import replace
 from typing import Any
 
 from haystack import tracing
@@ -18,36 +16,60 @@ from .span_records import (
     EvalCaseUsage,
     ReportedUsage,
     SpanRecord,
-    eval_case_usage_from_records,
-    get_component_name,
-    is_generator_span,
-    measure_output,
+    SpanRecords,
+    _eval_case_usage_from_records,
 )
 
 # The tags a component's output arrives under.
 USAGE_OUTPUT_TAGS = ("haystack.component.output", "haystack.agent.step.llm.output")
 
+# How many of a socket's strings to keep, and how much of each.
+MAX_RECORDED_TEXTS = 8
+MAX_RECORDED_TEXT_CHARS = 120
 
-@dataclass
-class SpanRecords:
+
+def _capped(text: str) -> str:
     """
-    Every span record collected under one eval case span.
+    Cut one recorded string to its allowance, marking the cut so a reader knows there was more.
 
-    :param records: The records, in the order their spans ended.
-    :param lock: Guards appends, since a run's components may be traced from several threads or tasks.
+    :param text: The string a component emitted.
+    :returns: The string, ending in an ellipsis when anything was dropped.
     """
+    return text if len(text) <= MAX_RECORDED_TEXT_CHARS else f"{text[:MAX_RECORDED_TEXT_CHARS]}..."
 
-    records: list[SpanRecord] = field(default_factory=list)
-    lock: LockType = field(default_factory=Lock, repr=False)
 
-    def add(self, record: SpanRecord) -> None:
-        """
-        Collect one closed span's record.
+def _measure_output(value: dict[str, Any]) -> tuple[dict[str, int], dict[str, list[str]]]:
+    """
+    Measure how much a component emitted on each socket, and sample whatever it emitted as short strings.
 
-        :param record: What the span reported.
-        """
-        with self.lock:
-            self.records.append(record)
+    :param value: The component's output, by socket.
+    :returns: How many items each socket carried, and a capped sample of the sockets carrying only strings.
+    """
+    sizes: dict[str, int] = {}
+    texts: dict[str, list[str]] = {}
+    for socket, emitted in value.items():
+        # A socket carrying one string or one object emits one item, not none: a router or a prompt builder
+        # belongs in the chain as much as a retriever does. A string is characters, not items, so it counts once.
+        items = list(emitted) if isinstance(emitted, (list, tuple)) else [emitted]
+        sizes[socket] = len(items)
+        # Documents and messages are counted and dropped, so nothing long is retained by accident.
+        if items and all(isinstance(item, str) for item in items):
+            texts[socket] = [_capped(text=item) for item in items[:MAX_RECORDED_TEXTS]]
+    return sizes, texts
+
+
+def _is_generator_span(operation_name: str, tags: dict[str, Any]) -> bool:
+    """
+    Decide whether a span is a model call, whose token usage is what it reports.
+
+    :param operation_name: The span's operation name.
+    :param tags: The span's tags.
+    :returns: True for an agent step, a component-held generator, or a generator run as a pipeline component.
+    """
+    return operation_name in ("haystack.agent.step.llm", "haystack.chat_generator.run") or (
+        operation_name == "haystack.component.run"
+        and str(tags.get("haystack.component.type", "")).endswith("ChatGenerator")
+    )
 
 
 class _HarnessSpan(Span):
@@ -56,7 +78,7 @@ class _HarnessSpan(Span):
         Create a span that reports into one eval case.
 
         :param collected: Where this span's record goes when it closes, or `None` when the span happened
-            outside any eval case and nothing it reports is kept.
+            outside any eval case, in which case its record is dropped instead.
         :param record: What this span will fill in as it runs.
         """
         self.collected = collected
@@ -67,23 +89,27 @@ class _HarnessSpan(Span):
 
     def set_content_tag(self, key: str, value: Any) -> None:
         """Measure one component output and discard it, so no content is retained and none has to be enabled."""
-        if self.collected is None or key not in USAGE_OUTPUT_TAGS or not isinstance(value, dict):
+        if key not in USAGE_OUTPUT_TAGS or not isinstance(value, dict):
             return
         if not self.record.is_generator_span:
             # A generator reports what it spent, not how much reached the next stage, so measuring its
             # replies would only produce a size the measurement throws away.
-            self.record.output_sizes, self.record.output_texts = measure_output(value=value)
+            sizes, texts = _measure_output(value=value)
+            self.record = replace(self.record, output_sizes=sizes, output_texts=texts)
         else:
-            self.record.reported_output = True
-            self.record.reported_usage = [
-                ReportedUsage(model=reply.meta.get("model"), tokens=reply.meta.get("usage") or {})
-                for reply in value.get("replies") or []
-            ]
+            self.record = replace(
+                self.record,
+                reported_output=True,
+                reported_usage=[
+                    ReportedUsage(model=reply.meta.get("model"), tokens=reply.meta.get("usage") or {})
+                    for reply in value.get("replies") or []
+                ],
+            )
 
 
 class HarnessTracer(Tracer):
     """
-    Record what every span under an eval case reported, for `eval_case_usage_from_records` to make a measurement of.
+    Record what every span under an eval case reported, for `_eval_case_usage_from_records` to make a measurement of.
 
     Three things are taken from the spans a run emits and nothing else is kept: the token usage a generator
     reports, how many items every other component emitted, and a capped sample of the sockets that emitted
@@ -107,8 +133,8 @@ class HarnessTracer(Tracer):
             collected=collected,
             record=SpanRecord(
                 parent_span_id=parent.record.span_id if parent is not None else None,
-                component_name=get_component_name(tags=tags or {}),
-                is_generator_span=is_generator_span(operation_name=operation_name, tags=tags or {}),
+                component_name=str((tags or {}).get("haystack.component.name") or "") or None,
+                is_generator_span=_is_generator_span(operation_name=operation_name, tags=tags or {}),
             ),
         )
         token = self._span.set(span)
@@ -134,7 +160,7 @@ class HarnessTracer(Tracer):
             tracing.disable_tracing()
 
 
-def usage_from_span(span: Span) -> EvalCaseUsage:
+def _eval_case_usage_from_span(span: Span) -> EvalCaseUsage:
     """
     Return what a harness measured under one eval case span.
 
@@ -142,5 +168,5 @@ def usage_from_span(span: Span) -> EvalCaseUsage:
     :returns: The measurement, or an empty one when a HarnessTracer was not the active tracer.
     """
     if isinstance(span, _HarnessSpan) and span.collected is not None:
-        return eval_case_usage_from_records(records=span.collected.records)
+        return _eval_case_usage_from_records(records=span.collected.records)
     return EvalCaseUsage()
